# hub/src/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Kanban Board API",
    "description": "Kanban board backend API with Better Auth and Socket.IO",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Development server"
    }
  ],
...
  "paths": {
...
    "/api/agents": {
      "get": {
        "tags": [
          "Agents"
        ],
        "summary": "List user's agents",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 500,
              "default": 200
            },
            "description": "Max agents to return (newest first). Totals reflect all agent keys for the user."
          }
        ],
        "responses": {
          "200": {
            "description": "Agents retrieved",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentListResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Failed to list agents",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "operationId": "listAgents"
      },
      "post": {
        "tags": [
          "Agents"
        ],
        "summary": "Create a new agent",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "name"
                ],
                "properties": {
                  "name": {
                    "type": "string"
                  },
                  "description": {
                    "type": "string"
                  },
                  "expiresIn": {
                    "type": "integer",
                    "description": "Expiration time in seconds (minimum 60)"
                  },
                  "avatarSlug": {
                    "type": "string",
                    "maxLength": 120,
                    "pattern": "^[a-zA-Z0-9._-]+$",
                    "description": "Optional built-in avatar id; `.svg` suffix is stripped server-side."
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Agent created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateAgentResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Failed to create agent",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "operationId": "createAgents"
      }
    },
    "/api/agents/{id}": {
      "patch": {
        "tags": [
          "Agents"
        ],
        "summary": "Update agent",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string"
                  },
                  "description": {
                    "type": "string"
                  },
                  "isActive": {
                    "type": "boolean"
                  },
                  "expiresIn": {
                    "type": "integer",
                    "description": "New expiration time in seconds (minimum 60)"
                  },
                  "avatarSlug": {
                    "oneOf": [
                      {
                        "type": "string",
                        "maxLength": 120,
                        "pattern": "^[a-zA-Z0-9._-]+$",
                        "description": "Set metadata.avatarSlug"
                      },
                      {
                        "type": "string",
                        "enum": [
                          ""
                        ],
                        "description": "Empty string removes avatarSlug from metadata"
                      }
                    ],
                    "description": "Omit to leave unchanged; valid slug or \"\" to clear."
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Agent updated",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "agent": {
                      "$ref": "#/components/schemas/Agent"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid parameters or body",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Agent not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Failed to update agent",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "operationId": "updateAgents"
      },
      "delete": {
        "tags": [
          "Agents"
        ],
        "summary": "Delete agent",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Agent deleted"
          },
          "400": {
            "description": "Invalid agent id parameter",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Agent not found"
          },
          "500": {
            "description": "Failed to delete agent",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "operationId": "deleteAgents"
      }
    },
    "/api/agents/{id}/regenerate-key": {
      "post": {
        "tags": [
          "Agents"
        ],
        "summary": "Regenerate agent API key",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Key regenerated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RegenerateKeyResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Agent not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Failed to regenerate key",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "operationId": "createAgentsRegeneratekey"
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Enter the JWT token received from login/register endpoints"
      },
      "agentAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "x-agent-api-key",
        "description": "Agent API key authentication using the x-agent-api-key header. Provide the raw key value (for example: ag_xxx). Agents must have active delegations to access projects."
      }
    },
    "schemas": {
      "ErrorResponse": {
        "type": "object",
        "description": "Standard error response",
        "properties": {
          "error": {
            "type": "string",
            "description": "Error type or category"
          },
          "message": {
            "type": "string",
            "description": "Human-readable error message"
          },
          "details": {
            "type": "object",
            "description": "Additional error details"
          }
        },
        "required": [
          "error"
        ]
      },
...
      "Agent": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "prefix": {
            "type": "string"
          },
          "isActive": {
            "type": "boolean"
          },
          "lastUsedAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "expiresAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time",
            "description": "Present on PATCH /api/agents/:id response"
          },
          "metadata": {
            "type": "object",
            "additionalProperties": true,
            "properties": {
              "isAgent": {
                "type": "boolean"
              },
              "description": {
                "type": "string",
                "nullable": true
              },
              "createdBy": {
                "type": "integer"
              },
              "avatarSlug": {
                "type": "string",
                "nullable": true,
                "description": "Built-in avatar id (no .svg)"
              }
            }
          }
        }
      },
      "CreateAgentResponse": {
        "type": "object",
        "properties": {
          "agent": {
            "$ref": "#/components/schemas/Agent"
          },
          "apiKey": {
            "type": "string"
          }
        }
      },
      "RegenerateKeyResponse": {
        "type": "object",
        "properties": {
          "agent": {
            "$ref": "#/components/schemas/Agent"
          },
          "apiKey": {
            "type": "string"
          }
        }
      },
      "AgentListResponse": {
        "type": "object",
        "properties": {
          "agents": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Agent"
            }
          },
          "total": {
            "type": "integer",
            "description": "Count of agent API keys for this user"
          },
          "activeTotal": {
            "type": "integer",
            "description": "Subset with enabled !== false"
          },
          "limit": {
            "type": "integer",
            "description": "Applied page size (query limit, clamped 1\u2013500, default 200)"
          }
        }
      },
...
    }
  }
}